requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
google-api-core==2.5.0
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==3.0.7
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.3.24
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.2